Retrieving network stats from proc path provided by env "PROCFS_MOUNT"
